refactor: removes weird non-lua syntax
removed the non-lua stuff from my old lovepack lua management thinger
script and instead converted over to real lua. had to add some stuff for
loading and checking dependencies though, not too sure if i like it
since it adds a lot of overhead to each script and fragments different
loader versions (which is what i was trying to get rid of by using my
weird non-lua stuff) but anyway, this makes it easier for people who
don't want lovepack or have any idea what it is, to still use the
library.

the only problem is they need to download all the dependencies, but the
readme will mention it and the actual library itself will mention it
when you run it.

on that note, modified the lua loader stuff for the busted test script
so it actually works and loads the stringtools dependency.

and also adds some extra code in the init.lua that looks for and loads
the stringtools library. also checks that it is the correct library and
not just something that is named the same. there is a version
requirement thing, but i don't want the script to check that because
that is A LOT of overhead (by that i mean i need to write code into the
library that isn't about the library ... and why would i do that?). this
will be checked and addressed if you use the new version of lovepack (in
development) and it will check these things when compiling libraries to
single files (in development). but anyways...

this only refactors that library code (the glue) and none of the actual
function, so i didn't increment the library at all, though probably
should have...