fix: fixes dependency loading issue
fixes an issue where you require this library not in the root of
your project, and it can't find the dependecies because it is looking
in the root of your project, not the folder you put this library.

adds code to break appart the require path `a.b.c.d.library` and pulls
out the parent folders `a.b.c.d.` and adds it to the dependency when it
loads it `a.b.c.d.dependency` so it is now in the right folder.